requirements.txt
absl-py==0.7.1
appnope==0.1.0
astor==0.7.1
backcall==0.1.0
bert-tensorflow==1.0.1
certifi==2019.3.9
decorator==4.4.0
gast==0.2.2
grpcio==1.16.1
h5py==2.9.0
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
mkl-fft==1.0.12
mkl-random==1.0.2
mock==2.0.0
numpy==1.16.3
parso==0.4.0
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
Pygments==2.4.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-hub==0.4.0
termcolor==1.1.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.2
